requirements.txt
# LLM providers
openai>=1.50.0
anthropic>=0.40.0
google-generativeai>=0.8.0

# Orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# RAG / Vector store
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Data & evaluation
requests>=2.32.0
scipy>=1.13.0
numpy>=1.26.0
pandas>=2.0.0
plotly>=5.20.0

# UI & config
streamlit>=1.40.0
supabase>=2.7.4
python-dotenv>=1.0.0
pydantic>=2.8.0
tiktoken>=0.7.0
